# api-spec.yaml
openapi: 3.0.1
info:
  title: Chattrix API
  description: |
    API toàn diện cho ứng dụng Chattrix, bao gồm xác thực, quản lý người dùng, và tính năng gọi điện thời gian thực.

    ---

    ### Chi tiết các sự kiện WebSocket
    Tất cả các tin nhắn từ server đều được bọc trong một đối tượng `WebSocketMessage` chung.
    
    **Message Events:**
    *   **`chat.message`**: Tin nhắn trò chuyện mới. (Payload: `OutgoingMessageDto`)
    *   **`message.updated`**: Một tin nhắn đã được cập nhật. (Payload: `MessageUpdateDto`)
    *   **`message.deleted`**: Một tin nhắn đã bị xóa. (Payload: `MessageDeleteDto`)
    *   **`message.reaction`**: Reaction đã được thêm/xóa khỏi tin nhắn. (Payload: `ReactionEventDto`)
    
    **Conversation Events:**
    *   **`conversation.update`**: Thông tin cuộc trò chuyện đã được cập nhật (ví dụ: tin nhắn cuối cùng). (Payload: `ConversationUpdateDto`)
    *   **`typing.indicator`**: User đang typing trong conversation. (Payload: `TypingIndicatorResponseDto`)
    
    **User Events:**
    *   **`user.status`**: Trạng thái trực tuyến của người dùng đã thay đổi. (Payload: `UserStatusDto`)
    
    **Friend Request Events:**
    *   **`friend.request.received`**: Nhận được friend request mới. (Payload: `FriendRequestResponse`)
    *   **`friend.request.accepted`**: Friend request được chấp nhận. (Payload: `FriendRequestResponse`)
    *   **`friend.request.rejected`**: Friend request bị từ chối. (Payload: `FriendRequestRejectDto`)
...
    
...
    
...
servers:
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    ErrorResponse:
      type: object
      required:
        - success
        - error
      properties:
        success:
          type: boolean
          example: false
        message:
          type: string
          description: "Only present for validation errors (e.g., 'Validation failed'). Single business errors don't have this field."
          example: "Validation failed"
        error:
          type: object
          required:
            - code
            - message
          properties:
            code:
              type: string
              description: "Error code (e.g., VALIDATION_ERROR, UNAUTHORIZED, USER_NOT_FOUND)"
              example: "RESOURCE_NOT_FOUND"
            message:
              type: string
              description: "Human-readable error message"
              example: "The requested resource was not found"
            details:
              type: object
              description: "Map of field names to error messages (only for validation errors with multiple field errors)"
              additionalProperties:
                type: string
              example:
                email: "Email already exists"
                username: "Username must be 4-20 characters"
        requestId:
          type: string
          description: "Unique request ID for tracking"
          example: "req-12345-abcde"
...
    FriendRequestResponse:
      type: object
      properties:
        id: { type: integer, format: int64 }
        userId: { type: integer, format: int64, description: "ID của người gửi/nhận (tùy context)" }
        username: { type: string }
        fullName: { type: string }
        avatarUrl: { type: string, format: uri, nullable: true }
        status: { type: string, enum: [PENDING, ACCEPTED, REJECTED, BLOCKED] }
        nickname: { type: string, nullable: true }
        online: { type: boolean }
        requestedAt: { type: string, format: date-time }
        acceptedAt: { type: string, format: date-time, nullable: true }
        rejectedAt: { type: string, format: date-time, nullable: true }
      example:
        id: 15
        userId: 20
        username: "user20"
        fullName: "Tran Thi B"
        avatarUrl: "https://res.cloudinary.com/dk3gud5kq/image/upload/v1765476687/avatars/20.png"
        status: "PENDING"
        nickname: null
        online: false
        requestedAt: "2025-12-14T07:00:00.000Z"
        acceptedAt: null
        rejectedAt: null
    SendFriendRequestRequest:
      type: object
      required: [receiverUserId]
      properties:
        receiverUserId:
          type: integer
          format: int64
          description: "User ID của người nhận friend request"
        nickname:
          type: string
          nullable: true
          description: "Nickname cho friend (optional)"

    # --- Read Receipts Schemas ---
...
security:
  - bearerAuth: []
...
paths:
...
  /v1/friend-requests:
    post:
      summary: "Send friend request"
      description: |
        Gửi lời mời kết bạn đến một người dùng.
        
        **Status Flow:**
        1. PENDING - Request được gửi, chờ accept/reject
        2. ACCEPTED - Request được chấp nhận, tạo bidirectional contact
        3. REJECTED - Request bị từ chối
        
        **Note:** Không thể gửi request nếu:
        - Đã có contact relationship (PENDING/ACCEPTED/BLOCKED)
        - Receiver là chính mình
      tags: [Social]
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/SendFriendRequestRequest' }
            example:
              receiverUserId: 30
              nickname: "Potential Friend"
      responses:
        '201':
          description: "Friend request sent successfully."
          content:
            application/json:
              schema:
                type: object
                properties:
                  success: { type: boolean, example: true }
                  message: { type: string }
                  data: { $ref: '#/components/schemas/FriendRequestResponse' }
              example:
                success: true
                message: "Friend request sent successfully"
                data:
                  id: 25
                  userId: 30
                  username: "user30"
                  fullName: "Hoang Van E"
                  avatarUrl: null
                  status: "PENDING"
                  nickname: "Potential Friend"
                  online: false
                  requestedAt: "2025-12-14T09:30:00.000Z"
                  acceptedAt: null
                  rejectedAt: null
        '400':
          description: "Bad Request - Validation failed."
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              examples:
                send_to_yourself:
                  summary: "Cannot send to yourself"
                  value:
                    success: false
                    error:
                      code: "BAD_REQUEST"
                      message: "Cannot send friend request to yourself"
                    requestId: "req-friendreq001"
                already_sent:
                  summary: "Already sent"
                  value:
                    success: false
                    error:
                      code: "BAD_REQUEST"
                      message: "Friend request already sent"
                    requestId: "req-friendreq002"
                already_friends:
                  summary: "Already friends"
                  value:
                    success: false
                    error:
                      code: "BAD_REQUEST"
                      message: "Already friends"
                    requestId: "req-friendreq003"
                user_blocked:
                  summary: "User is blocked"
                  value:
                    success: false
                    error:
                      code: "BAD_REQUEST"
                      message: "Cannot send friend request to blocked user"
                    requestId: "req-friendreq004"
        '401':
          description: "Unauthorized - Invalid or expired token."
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              example:
                success: false
                error:
                  code: "UNAUTHORIZED"
                  message: "Invalid or expired token"
                requestId: "req-friendreq005"
        '404':
          description: "User not found."
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              example:
                success: false
                error:
                  code: "RESOURCE_NOT_FOUND"
                  message: "Receiver not found"
                requestId: "req-friendreq006"
  /v1/friend-requests/received:
    get:
      summary: "Get pending friend requests received"
      description: |
        Lấy danh sách lời mời kết bạn mà user hiện tại đã nhận (chưa accept/reject).
        
        **Use case:** Hiển thị notification "X wants to be your friend"
      tags: [Social]
      security:
        - bearerAuth: []
      responses:
        '200':
          description: "Pending requests retrieved successfully."
          content:
            application/json:
              schema:
                type: object
                properties:
                  success: { type: boolean, example: true }
                  message: { type: string }
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/FriendRequestResponse'
              example:
                success: true
                message: "Pending requests retrieved successfully"
                data:
                  - id: 20
                    userId: 25
                    username: "user25"
                    fullName: "Pham Van D"
                    avatarUrl: "https://res.cloudinary.com/dk3gud5kq/image/upload/v1765476687/avatars/25.png"
                    status: "PENDING"
                    nickname: null
                    online: true
                    requestedAt: "2025-12-14T08:00:00.000Z"
                    acceptedAt: null
                    rejectedAt: null
                  - id: 22
                    userId: 28
                    username: "user28"
                    fullName: "Nguyen Thi F"
                    avatarUrl: null
                    status: "PENDING"
                    nickname: null
                    online: false
                    requestedAt: "2025-12-14T07:30:00.000Z"
                    acceptedAt: null
                    rejectedAt: null
        '401':
          description: "Unauthorized - Invalid or expired token."
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              example:
                success: false
                error:
                  code: "UNAUTHORIZED"
                  message: "Invalid or expired token"
                requestId: "req-friendreqrecv001"
  /v1/friend-requests/sent:
    get:
      summary: "Get pending friend requests sent"
      description: |
        Lấy danh sách lời mời kết bạn mà user hiện tại đã gửi (chưa được accept/reject).
        
        **Use case:** Hiển thị danh sách "Pending requests" để có thể cancel
      tags: [Social]
      security:
        - bearerAuth: []
      responses:
        '200':
          description: "Sent requests retrieved successfully."
          content:
            application/json:
              schema:
                type: object
                properties:
                  success: { type: boolean, example: true }
                  message: { type: string }
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/FriendRequestResponse'
              example:
                success: true
                message: "Sent requests retrieved successfully"
                data:
                  - id: 18
                    userId: 22
                    username: "user22"
                    fullName: "Tran Van G"
                    avatarUrl: null
                    status: "PENDING"
                    nickname: "Work Colleague"
                    online: false
                    requestedAt: "2025-12-14T06:00:00.000Z"
                    acceptedAt: null
                    rejectedAt: null
        '401':
          description: "Unauthorized - Invalid or expired token."
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              example:
                success: false
                error:
                  code: "UNAUTHORIZED"
                  message: "Invalid or expired token"
                requestId: "req-friendreqsent001"
  /v1/friend-requests/{requestId}/accept:
    post:
      summary: "Accept friend request"
      description: |
        Chấp nhận lời mời kết bạn.
        
        **Action:**
        - Set request status = ACCEPTED
        - Tạo reverse contact (bidirectional friendship)
        - Both users become friends
      tags: [Social]
      security:
        - bearerAuth: []
      parameters:
        - name: requestId
          in: path
          required: true
          schema: { type: integer, format: int64 }
          description: "ID của friend request cần accept"
      responses:
        '200':
          description: "Friend request accepted successfully."
          content:
            application/json:
              schema:
                type: object
                properties:
                  success: { type: boolean, example: true }
                  message: { type: string }
                  data: { $ref: '#/components/schemas/FriendRequestResponse' }
              example:
                success: true
                message: "Friend request accepted successfully"
                data:
                  id: 20
                  userId: 25
                  username: "user25"
                  fullName: "Pham Van D"
                  avatarUrl: "https://res.cloudinary.com/dk3gud5kq/image/upload/v1765476687/avatars/25.png"
                  status: "ACCEPTED"
                  nickname: null
                  online: true
                  requestedAt: "2025-12-14T08:00:00.000Z"
                  acceptedAt: "2025-12-14T09:45:00.000Z"
                  rejectedAt: null
        '400':
          description: "Bad Request - Validation failed."
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              examples:
                not_receiver:
                  summary: "Not the receiver"
                  value:
                    success: false
                    error:
                      code: "BAD_REQUEST"
                      message: "You are not the receiver of this request"
                    requestId: "req-acceptreq001"
                not_pending:
                  summary: "Request not pending"
                  value:
                    success: false
                    error:
                      code: "BAD_REQUEST"
                      message: "Friend request is not pending"
                    requestId: "req-acceptreq002"
        '401':
          description: "Unauthorized - Invalid or expired token."
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              example:
                success: false
                error:
                  code: "UNAUTHORIZED"
                  message: "Invalid or expired token"
                requestId: "req-acceptreq003"
        '404':
          description: "Request not found."
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              example:
                success: false
                error:
                  code: "RESOURCE_NOT_FOUND"
                  message: "Friend request not found"
                requestId: "req-acceptreq004"
  /v1/friend-requests/{requestId}/reject:
    post:
      summary: "Reject friend request"
      description: |
        Từ chối lời mời kết bạn.
        
        **Action:**
        - Set request status = REJECTED
        - Keep the record for history (not deleted)
      tags: [Social]
      security:
        - bearerAuth: []
      parameters:
        - name: requestId
          in: path
          required: true
          schema: { type: integer, format: int64 }
          description: "ID của friend request cần reject"
      responses:
        '204':
          description: "Friend request rejected successfully (no content)."
        '400':
          description: "Bad Request - Validation failed."
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              examples:
                not_receiver:
                  summary: "Not the receiver"
                  value:
                    success: false
                    error:
                      code: "BAD_REQUEST"
                      message: "You are not the receiver of this request"
                    requestId: "req-rejectreq001"
                not_pending:
                  summary: "Request not pending"
                  value:
                    success: false
                    error:
                      code: "BAD_REQUEST"
                      message: "Friend request is not pending"
                    requestId: "req-rejectreq002"
        '401':
          description: "Unauthorized - Invalid or expired token."
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              example:
                success: false
                error:
                  code: "UNAUTHORIZED"
                  message: "Invalid or expired token"
                requestId: "req-rejectreq003"
        '404':
          description: "Request not found."
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              example:
                success: false
                error:
                  code: "RESOURCE_NOT_FOUND"
                  message: "Friend request not found"
                requestId: "req-rejectreq004"
  /v1/friend-requests/{requestId}:
    delete:
      summary: "Cancel friend request"
      description: |
        Hủy lời mời kết bạn đã gửi (chỉ sender mới có thể cancel).
        
        **Action:**
        - Delete the request record completely
        - Only PENDING requests can be cancelled
      tags: [Social]
      security:
        - bearerAuth: []
      parameters:
        - name: requestId
          in: path
          required: true
          schema: { type: integer, format: int64 }
          description: "ID của friend request cần cancel"
      responses:
        '204':
          description: "Friend request cancelled successfully (no content)."
        '400':
          description: "Bad Request - Validation failed."
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              examples:
                not_sender:
                  summary: "Not the sender"
                  value:
                    success: false
                    error:
                      code: "BAD_REQUEST"
                      message: "You are not the sender of this request"
                    requestId: "req-cancelreq001"
                not_pending:
                  summary: "Request not pending"
                  value:
                    success: false
                    error:
                      code: "BAD_REQUEST"
                      message: "Friend request is not pending"
                    requestId: "req-cancelreq002"
        '401':
          description: "Unauthorized - Invalid or expired token."
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              example:
                success: false
                error:
                  code: "UNAUTHORIZED"
                  message: "Invalid or expired token"
                requestId: "req-cancelreq003"
        '404':
          description: "Request not found."
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              example:
                success: false
                error:
                  code: "RESOURCE_NOT_FOUND"
                  message: "Friend request not found"
                requestId: "req-cancelreq004"